Ensure cinder container udev line isn't duplicated
With the included version of the lxc module in Juno, the udev mount line
would have 2 spaces on either side of the = sign. While this worked, it
conflicted with upgrades, which inserted the same line, except with 1
space on either side of the = sign. Having two lines that were
functionally equivalent meant that the container cannot restart.

This patch ensures that the correct line is always present, so that when
an upgrade happens the udev mount entry is not duplicated.

Closes-Bug #1497681

Change-Id: I43957c78478e7f01be271576ca9fcf4ac7828f72